Suppress /analyze noise

+++ compress/compress.cpp
@@ -247,7 +247,7 @@ namespace
 
     inline bool IsErrorTooLarge(float f, float threshold)
     {
-        return (fabs(f) > threshold) != 0;
+        return (fabsf(f) > threshold) != 0;
     }
 
     inline bool SkipCompressCases(DXGI_FORMAT format, DXGI_FORMAT cformat, DWORD flags, bool gpu)
