+++ filter/mipmaps.cpp
@@ -323,12 +323,12 @@ namespace
 
     inline bool IsErrorTooSmall(float f, float threshold)
     {
-        return (fabs(f) < threshold) != 0;
+        return (fabsf(f) < threshold) != 0;
     }
 
     inline bool IsErrorTooLarge(float f, float threshold)
     {
-        return (fabs(f) > threshold) != 0;
+        return (fabsf(f) > threshold) != 0;
     }
 }
