+++ tex/convert.cpp
@@ -653,7 +653,7 @@ namespace
 
     inline bool IsErrorTooLarge(float f, float threshold)
     {
-        return (fabs(f) > threshold) != 0;
+        return (fabsf(f) > threshold) != 0;
     }
 
     inline bool SkipMonoCase(DXGI_FORMAT format)
