+++ tex/premulalpha.cpp
@@ -22,12 +22,12 @@ namespace
 {
     inline bool IsEqual(float f1, float f2)
     {
-        return (fabs(f1 - f2) < 0.000001) != 0;
+        return (fabsf(f1 - f2) < 0.000001f) != 0;
     }
 
     inline bool IsErrorTooLarge(float f, float threshold)
     {
-        return (fabs(f) > threshold) != 0;
+        return (fabsf(f) > threshold) != 0;
     }
 }
